Updating status/tweeting
We added a basic get function and a basic post functions to tweet on twitter. We also made a function to return our credentials all in the authenticator.js . I also added more comments on the page. then we got the secret tokens and stored them in the twitter credentials. In the index we built a route for the tweet  if it works it returns a 200 status if it doesn't it is a 418.